+++ src/pubtools/pulplib/_impl/client_pulp3/client.py
@@ -409,3 +409,19 @@ class Pulp3Client:
             params["fields"] = ",".join(fields)
         data = await self._request("GET", url, params=params)
         return data.get("results", []) if data else []
+
+    async def get_distribution(self, name: str) -> Optional[Dict[str, Any]]:
+        """Get a distribution by name.
+
+        Args:
+            name: Name of the distribution
+
+        Returns:
+            Distribution dictionary
+        """
+        out = None
+        url = "/distributions/rpm/rpm/"
+        data = await self._request("GET", url, params={"name": name})
+        if data:
+            out = data["results"][0] if data.get("count") else None
+        return out