+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `get_distribution()` method for pulp3 client
+
 ## [2.44.0] - 2026-06-24
 
 ### Added